+++ common/src/main/java/traben/solid_mobs/client/solidMobsClient.java
@@ -1,0 +1,8 @@
+package traben.solid_mobs.client;
+
+public class solidMobsClient {
+
+    public static boolean haveServerConfig = false;
+
+
+}