+++ fabric-like/src/main/java/traben/solid_mobs/fabriclike/fabricLikeMain.java
@@ -1,0 +1,9 @@
+package traben.solid_mobs.fabriclike;
+
+import traben.solid_mobs.solidMobsMain;
+
+public class fabricLikeMain {
+    public static void init() {
+        solidMobsMain.init();
+    }
+}