+++ fabric/src/main/java/traben/solid_mobs/fabric/SolidMobsCrossPlatformHelperImpl.java
@@ -1,0 +1,12 @@
+package traben.solid_mobs.fabric;
+
+import net.fabricmc.loader.api.FabricLoader;
+
+import java.nio.file.Path;
+
+public class SolidMobsCrossPlatformHelperImpl {
+
+    public static Path getConfigDirectory() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+}