+++ forge/src/main/java/traben/solid_mobs/forge/SolidMobsCrossPlatformHelperImpl.java
@@ -1,0 +1,12 @@
+package traben.solid_mobs.forge;
+
+import net.minecraftforge.fml.loading.FMLPaths;
+
+import java.nio.file.Path;
+
+public class SolidMobsCrossPlatformHelperImpl {
+
+    public static Path getConfigDirectory() {
+        return FMLPaths.CONFIGDIR.get();
+    }
+}