+++ quilt/src/main/java/traben/solid_mobs/fabric/SolidMobsCrossPlatformHelperImpl.java
@@ -1,0 +1,12 @@
+package traben.solid_mobs.fabric;
+
+import org.quiltmc.loader.api.QuiltLoader;
+
+import java.nio.file.Path;
+
+public class SolidMobsCrossPlatformHelperImpl {
+
+    public static Path getConfigDirectory() {
+        return QuiltLoader.getConfigDir();
+    }
+}